save at a none broken part
