+++ shift-scan/src/app/(routes)/dashboard/myTeam/[id]/employee/[employeeId]/employee-info.tsx
@@ -1,5 +1,4 @@
 import prisma from "@/lib/prisma";
-import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
 import { TitleBoxes } from "@/components/(reusable)/titleBoxes";
 import { Sections } from "@/components/(reusable)/sections";
 import { Contents } from "@/components/(reusable)/contents";
@@ -7,8 +6,8 @@ import { Forms } from "@/components/(reusable)/forms";
 import { Labels } from "@/components/(reusable)/labels";
 import { Inputs } from "@/components/(reusable)/inputs";
 
-export default async function employeeInfo({ params }: Params) {
-  const id = params.id;
+export default async function employeeInfo({ params }: { params: { employeeId: string } }) {
+  const id = params.employeeId;
   const employee = await prisma.users.findUnique({
     where: {
       id: id.toString(),
