+++ shift-scan/src/app/api/(MyTeam)/getCrewById/[crewId]/route.ts
@@ -1,0 +1,49 @@
+"use server";
+import { NextResponse } from "next/server";
+import prisma from "@/lib/prisma";
+import { auth } from "@/auth";
+
+export async function GET(request: Request, { params }: { params: { crewId: string } }) {
+
+  const session = await auth();
+  const userId = session?.user.id;
+
+  if (!userId) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+  const crewId  = Number(params.crewId);
+
+  try{
+
+  const crewMembers = await prisma.crewMembers.findMany({
+    where: { crewId: crewId },
+    include: {
+      crew: {
+        select: {
+          crewMembers: {
+            select: {
+              user: {
+                select: {
+                  id: true,
+                  firstName: true,
+                  lastName: true,
+                  image: true,
+                },
+              },
+            },
+          },
+        },
+      },
+    },
+  });
+
+  const crew = crewMembers[0].crew.crewMembers.map((member) => member.user);
+  
+  // Set Cache-Control header for caching
+  return NextResponse.json(crew);
+}
+  catch(error) {
+    console.error("Error fetching crew data:", error);
+    return NextResponse.json({ error: "Failed to fetch crew data" }, { status: 500 });
+  }
+}
