+++ shift-scan/src/app/api/getAllJobsites/route.tsx
@@ -1,4 +1,4 @@
-"use server";
+ "use server";
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 import { auth } from "@/auth";
