+++ shift-scan/src/components/(animations)/spinner.tsx
@@ -1,0 +1,7 @@
+export default function Spinner() {
+    return (
+        <div className="flex justify-center items-center h-screen">
+            <div className="animate-spin rounded-full h-32 w-32 border-b-2 border-red-700"></div>
+        </div>
+    );
+}
